Minor: Remove datafusion-core dev dependency from datafusion-sql
