+++ datafusion/sql/Cargo.toml
@@ -42,6 +42,3 @@ datafusion-common = { path = "../common", version = "15.0.0" }
 datafusion-expr = { path = "../expr", version = "15.0.0" }
 log = "^0.4"
 sqlparser = "0.28"
-
-[dev-dependencies]
-datafusion = { path = "../core" }
