+++ datafusion/sql/src/planner.rs
@@ -3301,14 +3301,11 @@ fn ensure_any_column_reference_is_unambiguous(
 
 #[cfg(test)]
 mod tests {
-    use datafusion::arrow::array::ArrayRef;
-    use datafusion::prelude::SessionContext;
     use std::any::Any;
 
     use sqlparser::dialect::{Dialect, GenericDialect, HiveDialect, MySqlDialect};
 
     use datafusion_common::assert_contains;
-    use datafusion_expr::{create_udaf, Accumulator, AggregateState, Volatility};
 
     use super::*;
 
@@ -5318,64 +5315,6 @@ mod tests {
         \n  WindowAggr: windowExpr=[[APPROXMEDIAN(orders.qty) PARTITION BY [orders.order_id] ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING]]\
         \n    TableScan: orders";
         quick_test(sql, expected);
-    }
-
-    #[test]
-    fn udaf_as_window_func() -> Result<()> {
-        #[derive(Debug)]
-        struct MyAccumulator;
-
-        impl Accumulator for MyAccumulator {
-            fn state(&self) -> Result<Vec<AggregateState>> {
-                unimplemented!()
-            }
-
-            fn update_batch(&mut self, _: &[ArrayRef]) -> Result<()> {
-                unimplemented!()
-            }
-
-            fn merge_batch(&mut self, _: &[ArrayRef]) -> Result<()> {
-                unimplemented!()
-            }
-
-            fn evaluate(&self) -> Result<ScalarValue> {
-                unimplemented!()
-            }
-
-            fn size(&self) -> usize {
-                unimplemented!()
-            }
-        }
-
-        let my_acc = create_udaf(
-            "my_acc",
-            DataType::Int32,
-            Arc::new(DataType::Int32),
-            Volatility::Immutable,
-            Arc::new(|_| Ok(Box::new(MyAccumulator))),
-            Arc::new(vec![DataType::Int32]),
-        );
-
-        let mut context = SessionContext::new();
-        context.register_table(
-            TableReference::Bare { table: "my_table" },
-            Arc::new(datafusion::datasource::empty::EmptyTable::new(Arc::new(
-                Schema::new(vec![
-                    Field::new("a", DataType::UInt32, false),
-                    Field::new("b", DataType::Int32, false),
-                ]),
-            ))),
-        )?;
-        context.register_udaf(my_acc);
-
-        let sql = "SELECT a, MY_ACC(b) OVER(PARTITION BY a) FROM my_table";
-        let expected = r#"Projection: my_table.a, AggregateUDF { name: "my_acc", signature: Signature { type_signature: Exact([Int32]), volatility: Immutable }, fun: "<FUNC>" }(my_table.b) PARTITION BY [my_table.a] ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING
-  WindowAggr: windowExpr=[[AggregateUDF { name: "my_acc", signature: Signature { type_signature: Exact([Int32]), volatility: Immutable }, fun: "<FUNC>" }(my_table.b) PARTITION BY [my_table.a] ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING]]
-    TableScan: my_table"#;
-
-        let plan = context.create_logical_plan(sql)?;
-        assert_eq!(format!("{:?}", plan), expected);
-        Ok(())
     }
 
     #[test]
